Separate JS pump into own file, export constants from C
As a step toward generality of the method, this moves the JavaScript
side of the event pump into its own file, and doesn't communicate with
any UI widgets directly.

To help keep names straight, the C side of the pump is called `c-pump`
and the JS side is called `js-pump`.  Since they both are used as
.js files, the convention used is to call the compiled c-pump
"c-pump.o.js" to indicate its compiled nature (.o is for object files
coming out of C compilers)

The JavaScript side is changed to use strings for the event IDs, but
it uses all-caps #define-style names to line up with the C.  On the
C side, it provides these constants to JavaScript via a bit of inline
JavaScript--in the hopes to centralize these definitions.