Move JavaScript and C pump code to web worker
This moves the JavaScript side of the event pump--as well as the C
code for the event pump and the emscripten heap/API--into a web worker.
As a result, the UI can stay responsive even if the C code is running.

There is currently no way for the UI thread to exchange information
with the C code while it is running.  The only way data can move back
and forth is via message processing that happens while _c_on_event()
is not running.  The hope is to use "emterpreter" to run C code in
"JavaScript simulated in JavaScript" for portions that wish to have
asynchronous access to influences from the GUI.