Reorganize/rename for clarity, set focus to textbox
This gets rid of the echo back to JavaScript when it posts an event
to the C.  It wasn't being used and sort of muddied things--as to
why the pump in JavaScript for processing C requests was appearing
to be taking events from JavaScript.

Renames c_on_event() to on_js_event() along with that clarification.

The browser was remembering what was in the text box on reload, and
this clears that.  Also sets the focus after disablement, which is
needed in Chrome but not Firefox.  (The text box is not going to be
around long, but it's easy to fix that so why not.)