Separate JS out of HTML file, rename js-pump.js => worker.js
It's not good to have large JavaScript files glued into an HTML page,
so as a step in progressing this experiment this goes ahead and moves
it out into a separate file.

The name "js-pump.js" was a little misleading, given that the message
posting routines are "queueRequestToJS" and "postEventToC"...if c-pump.c
is the recipient of "queueEventToC" it makes it seem like js-pump.js
would be the receipient of "postRequestToJS" when it's actually to the
GUI.  Saying C/JS vs GUI/Worker might be a bit misleading as well,
so that can be reconsidered as well.