Switch from TextEdit to contenteditable, integrate halting
This commit explores a bit more territory with using a list of
divs to implement the console.  The span the user is typing in is made
into a contentEditable span, which gives it a bit more leeway to
blend in and flow across lines with other content.

It adds a first test of integrating the Escape key for halting, and
also a speculative test of a mechanic which might be used to implement
the "magic undo" feature from Ren Garden--by using mutation observers,
and then if Ctrl-Z doesn't cause a mutation assume a "magic undo".

It corrects the windows line endings in gui.js, adds a css file, and
puts the strict mode annotation on the javascript.