GNU AGPLv3 License, README, some multiline work
This is just more general tinkering on the project, including changing
all the lines to end in semicolons.

(Repository will be relaunched fresh when it's out of the experimental
stage, so git commits are not being micro-managed at this point--hence
the sloppy versioning.  Just keeping it fluid for right now.)