Use promise .then() waterfall, integrate user native
This changes the worker from using a single XMLHttpRequest() to using
the fetch() API and a series of .then() Promise steps.  These steps
more clearly lay out the boot sequence...and include converting the
GUI's message post callback saying the DOM is ready and emscripten's
onRuntimeInitialized callback into promises.

It adds in a boot step of fetching a Rebol file and executing it.  This
file includes a user native.  Currently the native runs on the worker
and its only UI ability is to write to the console.log(), but this
is a first step of getting the feature working.