Move all libRebol to GUI thread, use JS-AWAITER
This transitions the preliminary experiment which placed the libRebol
code on the worker into one that puts it onto the GUI thread.  This is
because if only one thread is able to access the heap, it should be
the GUI thread--as otherwise you cannot mix libRebol code with any
useful display-oriented JavaScript code (as well as other functions
that are limited to the GUI thread)

The demo is shifted to have more control logic written in Rebol.
Routines that simulate synchronous I/O are accomplished via JS-AWAITER,
which like a JS-NATIVE has a body that is a text string of JavaScript,
but unlike a JS-NATIVE it does not return just because that body
finishes.  Instead, the body has to call the resolve() function, like
a promise does...at which point the implicit "await" frees up and
the calling Rebol code will continue running.  This allows synchronous
printing and input.

Several files related to the previous experiment are deleted, with the
build notes migrated into the main interpreter for the %emscripten.r
config file.