Sync with January 2019 Ren-C changes
The only thing that really changed in the API that affected the demo
was that BUILTIN-EXTENSIONS is now a native function and not an API
entry point.  This means that code that wants to pick and activate
extensions out of the executable doesn't have to be C code.  It's
an easy enough change to call that instead of rebBuiltinExtension().

PRINT now accepts newline as the only single character it will print,
in order to be convenient and clear...as there are few better ways
to say PRINT NEWLINE than saying exactly that.  Having to put a hyphen
was not as nice.  So this syncs the ReplPad PRINT to do that.

The biggest change affecting what was written so far is in commenting
philosophy.  Now comments in the Ren-C codebase that are on the same
line as things try to put TWO spaces from the source--whether that is
C or Rebol.  This makes a lone semicolon a more acceptable comment
character.  For details:

https://forum.rebol.info/t/the-beta-one-comment-concession/1028