Get latest commit from GitHub, fetch that version from S3
This changes to use the Travis-built libr3.js.  It queries GitHub for
the latest commit ID (the API is CORS enabled, so requests of GitHub
are legal).  Then it uses that ID to get the version off of S3.

Additionally, it fetches the replpad.reb Rebol file off of GitHub as
well--and uses Rebol to decode it.