Add support for .wast, .asm.tmp.js locateFile()
For reasons unknown, it seems emscripten-built files started asking
for .wast (webassembly text) and .temp.asm.js files, that should only
be built in the debug build.  They're not being built, and it doesn't
try to fetch them from the network, yet it still asks where they would
be if they were built for some reason.

It's good to have support for them, but still weird that it's asking
in non-debug builds.  Investigate when there's time.