Escape HTML when printing output
Sometimes the console wants to output HTML (e.g. with a bolded prompt
or in instructional text with links).  But other times it wants the
console to output things like a TAG! visibly and normally as part of
the output.  This adds a /HTML switch to PRINT, and it will escape
if you don't use it.