Run console extension instead of custom loop
This changes from using the improvised REPL loop from the initial
experiment to using the actual CONSOLE extension.  One downside of this
is that it loses the ability to do HTML skinning of the prompts
(bolds, coloring, etc.) since that is done through strings which go
through normal PRINT (hence are assumed to be escaped).  But that
points to a general issue that needs solving for console dialects.

On the plus side, all the customization features are included--and
improvements to the console on either the plain REPL or JavaScript
will benefit both!