Let rebStartup() initialize extensions
In switching rebPromise() to be entirely part of the JavaScript
extension (and not a general API service), its implementation requires
loading the JavaScript extension.  This changes rebStartup() for JS
to go ahead and load all the builtin extensions.

(Review limiting this to only initializing the JavaScript extension,
and making other builtins optional).