Persistent scroll bar positions, jump cursor when keypress
This tunes the scrolling mechanic to try and adhere to the user's
"scroll bar intent"...e.g. if they have the scroll bar pinned to the
bottom, that it should track new output.  But if they are scrolled to
any other position, the terminal should not "jump" or lose their
selection.

It's based on a somewhat maniacal trick from a StackOverflow answer,
which accomplishes this by making the browser think the bottom of the
scroll region is actually the top (as pinning to scroll position 0
is a fixed coordinate that stays there, whereas you can't set a scroll
position to a number and not have it become outdated as more data gets
added to the bottom)

https://stackoverflow.com/a/34345634

This reverses the mouse wheel scrolling direction, so JavaScript is
needed to overcome that.  While it would be nice to have a simpler
solution, everything else was either more invasive or did not work.

Also makes it so that pressing a printable key will jump down to
the console input.