Tinker some with fonts
This makes both system-printed and user input information in the
console use a fixed width font, but the system output is semibold.
The Inconsolata font is chosen due to its good distinction of I, l,
0, and O characters.

Noto Sans is chosen for note text (e.g. what will be tutorial text)
just to look a little different from the Arial that everyone is used
to, while still being easily readable and not too gimmicky.