Add `filler` so that REPL aligns to top
The flipping trick to make scroll bars work caused the content to be
aligned to the bottom when there weren't enough lines to fill the
vertical space.  This throws in a filler element that will expand
to take up as much remaning space as possible, which coaxes the
content to the top.