Add call to rebShutdown(), hijack of QUIT
There's value in keeping some kind of QUIT command so that in a debug
build, it can do checking that's done on QUIT (various states balancing
to zero, etc.)  @BrianOtto proposes it might be a good time to remind
people of some links to click on.

This adds the call, and shows how the QUIT function can be ADAPT-ed
with a bit of prelude code, that could write to the ReplPad.