Make READ work for hosted TEXT! and GitHub blob URL!s
As a general rule, a web application is not allowed to make arbitrary
requests of domains besides the one they are hosted on.  The exception
is apis that use the "CORS" protocol (or workarounds with complicit
APIs that use "JSONP" and mask their data responses as fetched scripts)

GitHub offers a CORS API for its "blobs", which gives back a page's
data as Base64.  This recognizes that pattern when given to READ by
parsing the URL, and makes the appropriate fetch() request of GitHub.

This also handles a request for a FILE! as if it is a fetch().  This
only implements textual fetches, as the JavaScript API needs a design
for how reb.Binary() will handle array buffers.