Use replpad.reb directly from host, not hostilefork/master
It was not known initially how easy it would be for people to make
their own experimental deployments of the replpad to work on.  So one
idea for making it easier was that the script would be pulled
dynamically from a GitHub location using the CORS API.  Then perhaps
a URL fragment or other directive could point the system to their own
fork of the repository--but still using the library from the hosting
site.

Being able to create one's own ReplPad instance has proven to be no
particular barrier.  Additionally, the feature of pulling from GitHub
and decoding the Base64 via Rebol into a BINARY! has been moved into
READ, which could make it possible to reskin or re-load the console
after the fact anyway.