Simple floating dialog: do %webvid.reb
This takes a sample floating dialog with minimal dependencies from
the Internet and puts it into the page:

https://codepen.io/tovic/pen/XJEONy

It is a means to explore issues of message pump interaction with the
console.  Trying to build such systems out of pure JavaScript may be
a good idea or a bad idea.  (Probably a bad idea and better to use a
framework.)