Incorporate change to JS-AWAITER returning promise/value
With JS-AWAITER altered to use async functions under the hood, this
made it possible to hide the resolve/reject mechanics and instead
just return promises.  Also, if a non-promise value is returned from
an async function, that will be wrapped in a promise automatically.

This makes it slightly more difficut to work with classical callback
functions (like setTimeout) because the resolve/reject pair have to
be made with new Promise().  However, the rationale for needing the
several `setTimeout(resolve, 0)` instances doesn't apply with the
switch to returning promises, there always must be an asynchronous
yield to process the promise result anyway (the .next() never runs
immediately).