Move webvid demo to own directory, rename popupdemo
This particular demo is unlikely to be the basis of growing into any
serious VID effort.  So it was given a different name and moved out
of the root directory.

Since this makes it hard to find, it adopts the TAG! convention used
by DO to identify source by a memoized name.