Add stub for doing a top-level API test
A formal process or test project for the API is needed.  As the loader
is factored out so there is more than one libRebol.js-based page in
the world, it will be possible for a basic test site to exist
separately from the ReplPad project.

This is a simple beginning of a place to start adding tests in the
anticipation of that project.