Remove API-based GitHub file access from "blobs"
It actually turns out that raw.github links have CORS support.  This
makes the inefficient API access unnecessary.

However, support for making BINARY! is still needed.  In the interim,
this allows a READ that expects to convert binary to text to work by
inefficiently reading as text, and then converting to binary (such a
conversion will not be inefficient after UTF-8 everywhere).