Improve consistency of single-quoted string use in JS
There are not generally hard-and-fast rules for when to use single
quoted or double quoted strings in JavaScript.  But it seems that
mixing them arbitrarily is not ideal.

This attempts to use a guideline that if the material in question could
(pretty much) never have spaces/quotes/apostrophes in the slot it
represents, then single quotes works well.  This matches up somewhat
with the idea of a use of a "apostrophe'd WORD!" in Rebol--e.g. places
in JavaScript where Rebol would probably use a WORD! datatype.

Strings with spaces thus always default to using double quotes, unless
they contain double quotes themselvs--at which point it's a subjective
call on what to do.