READ GitLab "raw" style links via their CORS API
GitLab does not offer CORS access to its plain raw URLs for files in
a repository.  However, there is a transformation you can do to ask
for the file via their API which does offer the file as CORS.  This
injects that transformation into the READ layer of the REPL for
user convenience.