GitLab raw in JS-DO/CSS-DO, add /AUTOMIME, remove READ/STRING
This adds the capability to use GitLab raw files in JS-DO and CSS-DO.
In order to do so, it not only has to transform the URL into a CORS
compatible API call...but it also can't include the files directly
using the normal `<script src="...">` conventions, because GitLab
serves the files as `text/plain`:

https://stackoverflow.com/q/46593792

It is possible to subvert this by instead using the fetch() API, and
then handling the result as if it were eval()'d text or literal CSS.
That mechanism is added as a convenience on JS-DO and CSS-DO via a
refinement named /AUTOMIME (better name suggestions welcome).

This also removes the /STRING refinement from READ.  In addition to the
name being outdated, the idea that low-level READ would deal in
anything besides BINARY! is in question.  Design will hopefully follow
a copy of an existing successful model or a thought-out new one, vs.
trying to address cases with ad-hoc refinements.