Improve relative path handling (including CSS-DO, JS-DO)
Rebol had some logic by which WHAT-DIR was incorporated into the
concept of relative paths (so that DO of a FILE! in a web page would
be based on the idea that WHAT-DIR could actually be a URL! and that
the file would be looked at relative to that).

This brings that logic forward to the web by making the WHAT-DIR
equal to the browser's URL when the system starts up.  Then in each
script it will override that while the script is running.

Also adds handling to JS-DO and CSS-DO to incorporate WHAT-DIR into
their handling of FILE! as a relative path.