Change way of jumping the cursor on a printable keypress
Uses "deprecated" method of keypress event, but no suitable replacement
has emerged.  It will likely be supported in browsers for the
indefinite future, especially given the lack of a better way.

This makes it possible to use Ctrl-C, which was previously jumping the
cursor to the input if you had a selection since C is a printable
character.