Add JS-HEAD operation
Add operation which takes a URL and returns the HTTP headers as an
OBJECT! mapping keys to values.

It is presumed that using SET-WORD!s as keys in the object makes sense,
as this is one area where the web standards follow Rebol's core
notation.  Weighing the difference between MAP! and OBJECT! in this
is an ongoing quetion.  An object is chosen for now, because MAP!
does not (currently) consider x/y to be a match for SET-WORD!s.