Add notice for emterpreter users
This is a first stab at putting in a detection and message for
emterpreter users to point out that they should enable WASM threads.

It points to a GitHub wiki for starters so the instructions can be
gathered.  But it should ideally point to a nicer page at a nicer URL.