Move watchlist into its own component
The initial demo of ReplPad was a test of pulling together a few basic
non-jQuery-based widgets, to see if they could accomplish some of the
same functionality as Ren Garden, but in the browser.

A splitter, a table, and a right click menu were included.  However,
these aren't necessarily the best choices for the code, and might
distract from developing a reusable common console widget.

This factors the watchlist out, so it can be run with `do <watchlist>`
and still see the (non-functional) demo.  This should make it easier
to develop that separately, without polluting the ReplPad with another
experiment.