Resurrect "cool" scroll bar
It's possible to customize the scroll bar with CSS (in Chrome).
ReplPad's scroller had been made thinner, which was a neat effect...but
it broke at some point and went unnoticed.  (It was being applied to
the #replpad which under the new structure does not have a scroll bar,
so the style has to be applied to the parent.)

This restores the styling.