Use load_r3() promise, move WHAT-DIR setup to %load-r3.js
The code for establishing the working directory as the location in the
URL bar where the script starts up is now taken care of in %load-r3.js
Hence ReplPad can now take that for granted.

Also, the loading promise has gone through a few names.  Right now
it is `load_r3" (which won't be its final name, but it will likely
be a promise of this shape).