Deferred load for watchlist extension
The watchlist extension loads several files and does some processing,
so it should only be loaded if the user asks for it.  This helps make
the plain ReplPad boot faster.

Thie exposes the fact that the page layout is dependent on the
split.css file to get padding.  Pending that being sorted out, the
CSS file for split.css is included in the index.html from the
watchlist subdirectory