Make JS-DO run in global scope by default
Running JavaScript requires proxying to the GUI thread, and the logic
for doing that is all handled by JS-NATIVE.  If it were factored out
better it would be possible to reuse that code and do a lighter weight
form of it with JS-DO...but for the near term, it was just making a
function which took no arguments and calling it with EVAL.

This means the code was running in a function scope, so things like
a `var` declaration (for instance) would not be globally visible.
Sometimes you want global visibility, and it is also strictly more
powerful than local (since you can always wrap code in a function
and evaluate it).

Hence this changes the default for JS-DO to be global.  Rather than
make JS-DO its own native at this time, it achieves the global eval
by putting the text into a <script> tag.