Customize the console greeting to offer a REDBOL option
This adds the fledgling Redbol emulation to the web demo.  The way it
is done module-wise is not currently ideal, and it is at a very
early state.  But it now can at least somewhat run two Rebol2-ish
codebases, so probably time to show that it's in the works.