Use ordinary SYS/EXPORT for exporting overrides
With WAIT being removed from LIB in the JavaScript build (due to not
building the EVENT! extension), `lib/wait: ...` as an assignment would
cause an error.  It seemed that just adding WAIT into SYS/EXPORT would
not only add it to the user context, but also to LIB...and all the
definitions which were being mirrored into user and lib had the
same effect.

How modules work needs a thorough overall vetting--as most of what's
happened on them since R3-Alpha is maintenance to keep them running,
vs. actually critiquing the design.  But without knowing any better,
this seems to make exporting work for the time being.