Native JavaScript implementation of NOW
The core is now factoring several bits of "dead weight" for the
JavaScript build out into extensions (EVENT! code, local filesystem
and directory handling, Berkeley sockets networking...)

One of the pieces that was part of the "Host" abstraction was getting
the time itself.  This is also moved into an extension, which provides
the Rebol NOW function.

While it would be possible for JavaScript to use this extension, it
would mean also using the Emscripten emulation of the time APIs for
POSIX.  That's additional clutter, which can be expressed more
cleanly (and hopefully more optimally) by speaking directly to the
JavaScript APIs that those emulations must speak to.  This goes ahead
and does that, to not use the Time extension but rather to implement
NOW directly.

Time Zones are not currently accounted for.