Handle a PRINT immediately after a REPLPAD-RESET
The initial code established an empty "line" `<div>` to start the
printing process.  This was omitted when it produced an empty line
above the "note" `<div>`.  So the empty line was taken out, which
then disrupted the situation if you *didn't* have a note.

This changes it to make the "line" `<div>` on-demand, making it
possible to PRINT from after a REPLPAD-RESET.