Add JS-EVAL for getting back INTEGER!, TEXT! easily
In the general case, JS-DO cannot return a result--because it may
not have one.  (Perhaps it should be called JS-ELIDE?  :-/)

This adds a helper JS-EVAL for evaluations of single expressions that
can take place inside of a function.  It can be used to easily extract
a TEXT! or an INTEGER!.