Allow the DO-ing of scripts specified in URL bar
The "search string" is the part of the URL that comes after the
question mark (including the question mark, apparently).  This adds
a simple first crack at running scripts named by `?do=scriptname`,
using the `<tag>` running mode of DO.

Further design review and thinking on this feature has been requested,
but this is more-than-nothing to get started with.