First actual watch-ing WATCH
This fleshes out the WATCH feature a little more to actually do a
watch.  It does not have the dialecting features of Ren Garden (e.g.
numbers to fetch, delete, alias watches), nor does it evaluate code
(which would have to be tied in with cancellation, so there needs to
be some kind of cancel button on the watches).