Watchlist uses pure CSS, eliminate %table-resize.js
This uses `position: sticky` and `resize: horizontal` to achive a way
to resize columns in the watchlist in CSS without having to intercept
mouse moves and drags in JavaScript.

It doesn't have the conventional ability to drag via a splitter handle
between columns.  Instead, there's a small corner on the headers you
can drag with.  However, this is a luxury feature which is nowhere
close to the project's goals to worry about--so using a built-in
facility that is good enough is fine.

Since the watchlist is now loaded on demand for first usage, this
simplifies the initialization so that it assumes it is showing the
watchlist.  The ability to do JavaScript evals and pass in Rebol
values allows JavaScript helper routines to be moved into the
watchlist.js.