Changes to use reb.Startup() instead of load_r3
The `load_r3` export from %load-r3.js was changed from a Promise to
a function that returns a promise...with the name `reb.Startup()`.
This takes a configuration object that can specify callbacks to be
used in place of `console.log()`, warn(), error(), info()...but
could be extended to take other options in the future.

This allows some of the "replpad-aware" logging code to be removed
from %load-r3.js.