Wrap long lines in REPL
This switches the div to use `word-break: break-word;`, and has several
changes that are related to getting this to work.

One is to correctly split each line being added into its own div
(this had been broken at some point).

Another is to change from using `&nbsp;` escaping to the very helpful
CSS `white-space: pre-wrap;`.

Attempts were made at implementing a "hanging-indent" (so every
"virtual" line that is a continuation of the previous line is indented
a bit).  It didn't work, so that will have to be added later.