Guided Interactivity Test
This implements a sort of checklist for issues regarding the UI
behaviors of the REPL.  It is implemented by means of a function
called OKAY that gives some instructions and can manipulate the REPL.
It returns control to the user to use the REPL as normal each time,
but it remembers which step it is on, so that the next OKAY can do
the next step.

If a step fails, they can type NOPE and it will guide them to GitHub
issues (if there is an existing issue on the topic) to describe the
regression.

It depends on the recent change of wrapping long lines in the REPL.