Make it possible to run guided test with `?do=test-repl`
The guided test had code in it that expected a console to be available
in order to manipulate the prompt.  It tried to grab this console's
print-prompt function as it ran the first time, in order to be able
to restore it after each test step...if the test step changed it.

However, when you use the `?do=` feature it is like running code from
the command line.  There is no system/console object at all, because
the script is getting control.

Exactly how this might work in terms of scripts that want to deal with
the console needs to be rethought (maybe they'd just run `console`?)
In any case, this changes it so that it moves the logic for saving
and restoring the prompt into OKAY instead, and falls through when
the module runs to the console.