Switch reb.Did() to reb.DidQ()
Current thinking is that refinements may become NULL when unused.  If
that is the case, the lack of a quote on the reb.Did(reb.ArgR()) calls
would be a problem, because nulls cannot be evaluated.

This preventively switches to reb.DidQ().  Also changes the watchlist
to use the new-meaning of QUOTE instead of UNEVAL.