Use natural resolve() calls due to Asyncify benefits
While a JS-AWAITER was waiting for a resolve() call, the Rebol stack
had to be suspended.  JavaScript code would then run.  If you wanted
to do something like `resolve(reb.Text(js_string))`, it would mean
you would be trying to call more Rebol code as part of that resolve,
while the stack was still suspended.

This wasn't a problem for the pthreads build...because it just meant
WASM code would run on the GUI thread instead.  As long as you did not
try to suspend again you would be okay (and APIs like reb.Text() that
just construct values vs. running Rebol code should be in the clear).

However, the Emterpreter build was not designed to do this.  More
emulated WASM could not be added onto the stack (or start a new
emterpreter).  As a hack, the special response to doing a reb.Promise's
resolve() with a JavaScript function would call that function *after*
the resumption, so it could construct values at that time.

The nature of Asyncify resolves this problem, making the workaround
no longer necessary!