READ-STDIN rename to READ-LINE to match ASK
The ASK mezzanine is built on a lower-level I/O operation.  That
operation was previously called READ-STDIN.

However, the problem is that if READ-STDIN was assumed to be an alias
for the eventual concept of `READ STDIN` with that being an actual
PORT!, then it would not necessarily imply the same semantics as the
console...as STDIN has to be an agnostic byte-level service in order
to carry the full bandwidth of pipes (e.g. with CGI).  This would not
guarantee UTF-8, and have several other potential pitfalls.

To try and help move the meanings toward something more coherent, as
the consoles are unified (and become per-keypress-capable) the loaded
term of READ-STDIN was changed to a more prescriptive READ-LINE.
Unfortunately, the ReplPad had managed to "splice" behavior into ASK
by overriding READ-STDIN.  When it did not, this gave a puzzling
error.

Renaming the override helps from falling through to the default
READ-LINE behavior, restoring functionality to the console.