Add first attempt at automated browser "smoke test"
This adds a Python script that uses the "Marionette" API for controlling
a Firefox browser, to try and prove that the REPL can at least load
and print something.

It's a minimal first step toward establishing a more broad automated
testing answer for Rebol web apps (which could be tested with pretty
much any web testing framework).

This test is currently intended to be run using the `-headless` mode
of Firefox when it is enabled with `-marionette`, in Travis whenever
there is a change to the core interpreter.  It would be good to run it
on changes to the REPL as well (though usually it is assumed that such
changes are being tested, as the change is being made to the REPL!)