Redo escaping logic as Rebol code, add post-process
This makes URL links show up as clickable in console output, as well
as a little tweak for another demo that makes card suit unicode
characters show up in black and red.  These should be some kind of
hookable thing in the console, but for now it's relatively harmless
to have them.