Allow DO of GitLab and GitHub "decorated" links
Some URLs that represent executable code have a HTML presentation layer
on them.  This is why a GitHub link has a "raw" offering without all
that extra stuff on it (line numbers, buttons, etc.)

We don't want to hook at the READ level to redirect those UI pages to
give back the raw data...because you might want to READ and process the
UI decorations!  But if you ask to DO such a page, it's reasonable to
assume that what you actually wanted was to DO the raw content implied
by it.

This performs that forwarding for GitLab and GitHub UI links.