Fix DOWNLOAD
The download code hadn't been adjusted for NULL-refinements, and it
also was accidentally returning `foo.txt` instead of `%foo.txt` if
the data was textual.  This had been legal to use reb.Spell() on
when it was a WORD!, but once that became a TUPLE! it was no longer
legal.

It's probably best if TUPLE! does not blindly give back a textual
representation, as it may contain GROUP!s and this helps to force a
decision regarding what should be done with it.  So this just changes
to use a FILE!.