DOWNLOAD: MIME type handling, mandatory filename
To make DOWNLOAD "easy to use" it was originally done with a skippable
parameter for the filename.  But this just made it more confusing, so
this makes the filename mandatory--and only accepts FILE!

It also adds the ability to specify a MIME type, since that is
trivial to implement.  It defaults to text/plain or octet/stream based
on whether the data is TEXT! or BINARY!.