Fix `write clipboard://`
The stub for write was only set up to handle one thing, the clipboard.
It did this using the convention of when PARSE would match partial
inputs by default, so `parse clipboard:// ["clipboard:"]` would have
worked at that time.

That's a really bad way of doing it, and it also doesn't work now that
PARSE is back to insisting a full match unless the /PROGRESS option
is used with it.

A better way for doing this is schemes (which @rgchris is working on),
but this just goes ahead and explicitly checks for the URL! of
`clipboard://` just to demonstrate the test case working.