Handle NOW/TIME/PRECISE, correct nanoseconds value
Browsers offer a date getMilliseconds() option.  This was being
multiplied in NOW/PRECISE by 1000 instead of by a million to get
nanoseconds.

In addition to fixing that point, this also adds handling for the
case of NOW/TIME/PRECISE, which gets a more granular form of the
time.  Due to current limitations of MAKE TIME! (which need discussion
before canonizing as a general design), this is handled with a new
entry point specific for use in the implementation of NOW, called
MAKE-TIME-SN.