Clean up Storage support functions; Add File/Directory Schemes
The storage support functions no longer handle implicit storage type detection. This detection was based on the File schemes and should have more versatility.

The file schemes are possibly too smart for their own good. It is likely they could be made more robust by being more basic, particularly the directory scheme—left as an exercise for another time/place/person. All files are first encoded to Base64—this uses one third of the available storage space that is already sparse. Some space could be saved with a Base85 encoder, but the savings are marginal. For more space, a compatible filesystem could be rebuilt using the IndexedDB API—left as an exercise, etc...