Factor out %js-css-interop.reb and %storage.reb
This starts a process of reorganizing the replpad so that its functions
can live in different files.  (While running different files does add
overhead, we're not concerned about that at this time.)

In order to do this factoring, ReplPad is organized into stages that
first establish basic IO (to the JS console and to the ReplPad) for
getting debug output during load.  Then, it establishes the HTTP scheme
which is built on top of JavaScript's fetch().  Once this is done, it
can DO the URLs of files to provide additional functionality.

The introduction of the storage scheme disrupted the handling of FILE!
so that calls like `css-do %resources/whatever.css` are not working,
but this is pending resolution.