Restore Gitlab URL functionality
GitLab does not offer CORS on its raw links.  However, it does offer
CORS when data is accessed through the API.  So raw links have to be
turned into API accesses.

When the port changes were done, this functionality was moved into the
port layer.  However, it was commented out.

This removes that comment.  Also, since this is now handled at the
port level, it means that URLs do not have to be transformed to the
API URLs in other places like CSS-DO or JS-DO.  However, those places
do need to do translation to "raw" style links if given a decorated
gist or page with line numbers/etc.

(We don't want that kind of change to be happening at the READ layer,
because you may want to actually READ the version of the page that has
the decorations.)