SYM-XXX! => META-XXX!
Ultimately the JS dialect probably wants to use @xxx instead of ^xxx
to denote things to substitute via handle.  However, there are
currently issues trying to get that to fit in the number of types
when typesets are 64 bits.

Until that is remedied, this has to adapt the dialect to the new
META-XXX! names for the ^xxx forms.