JOIN and JOIN-ALL adjustments
JOIN became non-reducing of block arguments, which didn't affect the
uses here...but it was noticed that there were instances of JOIN COPY
which is redundant, as JOIN always creates a new series.

JOIN-ALL is no longer a thing.  (join datatype! [...]) is a better
solution as it tells you what type of a result you are expecting.