Move non-LIB commands to user, put useful ones in lib
This makes a small step toward the long goal of factoring the ReplPad
into a generic "Web app" library vs. the specific one app of the
console itself.

Functions like ABOUT and REDBOL are moved into the main.reb and
quarantined to the user context, while implementations of basic
platform services are exported to lib for other modules and scripts
to build on.