Add LATEST-OF to web build by default
The LATEST-OF command is a bit of a pain to run beyond just
`do <latest-of>` because it takes arguments.

This adds LATEST-OF as a default Web Repl command.

It would be nice if it could be cached and only load the script if
LATEST-OF is actually run, but some technical issues
with modules mean that once LATEST-OF is exported
the recipient (e.g. user context) will not see changes.  This is being
thought about, so it just runs the script for now.

Note that the CI runs LATEST-OF and ensures it gives back a URL! to
keep this functionality working.