Switch to @ form for doing named libraries
There's some contention over if `do @hostilefork` should act the same
as if you had code saying `if condition @hostilefork` or not.  But the
TAG! notation is definitely needed for a full string, because a form is
needed that has dots and slashes in it.  So moving in that direction.