Separate out testing steps to separate executions
Initially the browser was only instantiated once, to try and make
the build verification go faster.  But it's not really all that slow to
run in separate steps, and this gives better information about
what errored...as well as testing the ability of the action to run
several times in sequence (which seems to need work).